Move Fragments into view.main package
