+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
             android:name=".view.auth.AuthActivity"
             android:exported="false" />
         <activity
-            android:name=".view.MainActivity"
+            android:name=".view.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
