+++ app/src/main/res/layout/fragment_calendar.xml
@@ -9,9 +9,9 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".CalendarFragment">
+        tools:context=".view.main.CalendarFragment">
 
-        
+
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
