+++ app/src/main/res/layout/fragment_diary.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".DiaryFragment">
+        tools:context=".view.main.DiaryFragment">
